+++ bin/.gitignore
@@ -23,3 +23,10 @@
 /GameOverFrame$1.class
 /GameOverFrame.class
 /TestUI$1.class
+/PlayGame$1.class
+/PlayGame$2.class
+/PlayGame$3.class
+/PlayGame$4.class
+/PlayGame$5.class
+/PlayGame.class
+/Main.class
