+++ src/Main.java
@@ -1,10 +1,9 @@
-
-public class Test {
+public class Main {
 
 	public static void main(String[] args) {
-		
+
 		MainMenu mm = new MainMenu();
-		
+
 	}
 
 }
