+++ src/Shop.java
@@ -7,7 +7,7 @@ import javax.swing.JFrame;
 import javax.swing.JLabel;
 import java.awt.Color;
 
-public class Shop extends JFrame  {
+public class Shop extends JFrame {
 	private JButton use1;
 	private JButton use2;
 	private JButton use3;
@@ -16,20 +16,14 @@ public class Shop extends JFrame  {
 	private JButton buy2;
 	private JButton buy3;
 	private JButton buy4;
-	static String b1IsBought="false";
-    static String b2IsBought="false";
-	static String b3IsBought="false";
-	static String b4IsBought="false";
-	static int flag1=0;
-
-	
+	static String b1IsBought = "false";
+	static String b2IsBought = "false";
+	static String b3IsBought = "false";
+	static String b4IsBought = "false";
+	static int flag1 = 0;
 
 	static int life;
 
-	
-	
-	
-	
 	private void checkPrice(JLabel points) {
 		if (Integer.valueOf(points.getText()) < 100) {
 			buy1.setEnabled(false);
@@ -44,19 +38,20 @@ public class Shop extends JFrame  {
 			buy4.setEnabled(false);
 		}
 	}
-	
-	
 
 	public Shop() {
+		init();
+	}
+
+	private void init() {
 		getContentPane().setBackground(Color.WHITE);
 		setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-		
+
 		JLabel points = new JLabel("0");
 		points.setBounds(105, 588, 46, 14);
-		points.setText(String.valueOf(Integer.valueOf(Integer.valueOf(points.getText())
-				+ Integer.valueOf(TestUI.points))));
+		points.setText(String.valueOf(Integer.valueOf(Integer.valueOf(points
+				.getText()) + Integer.valueOf(PlayGame.points))));
 		getContentPane().add(points);
-		
 
 		this.setSize(900, 700);
 		this.setLocationRelativeTo(null);
@@ -96,17 +91,17 @@ public class Shop extends JFrame  {
 		if (Integer.valueOf(points.getText()) < 100) {
 			buy1.setEnabled(false);
 		}
-		
+
 		buy1.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				use1.setVisible(true);
-				TestUI.points = Integer.valueOf(points.getText()) - 100;
+				PlayGame.points = Integer.valueOf(points.getText()) - 100;
 				points.setText(String.valueOf(Integer.valueOf(points.getText()) - 100));
 				buy1.setEnabled(false);
 				checkPrice(points);
-				b1IsBought="true";
+				b1IsBought = "true";
 				MainMenu.saveData();
-				
+
 			}
 
 		});
@@ -119,13 +114,13 @@ public class Shop extends JFrame  {
 		}
 		buy2.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
-				
+
 				use2.setVisible(true);
-				TestUI.points = Integer.valueOf(points.getText()) - 250;
+				PlayGame.points = Integer.valueOf(points.getText()) - 250;
 				points.setText(String.valueOf(Integer.valueOf(points.getText()) - 250));
 				buy2.setEnabled(false);
 				checkPrice(points);
-				b2IsBought="true";
+				b2IsBought = "true";
 				MainMenu.saveData();
 			}
 		});
@@ -140,11 +135,11 @@ public class Shop extends JFrame  {
 			public void actionPerformed(ActionEvent e) {
 				checkPrice(points);
 				use3.setVisible(true);
-				TestUI.points = Integer.valueOf(points.getText()) - 500;
+				PlayGame.points = Integer.valueOf(points.getText()) - 500;
 				points.setText(String.valueOf(Integer.valueOf(points.getText()) - 500));
 				buy3.setEnabled(false);
 				checkPrice(points);
-				b3IsBought="true";
+				b3IsBought = "true";
 				MainMenu.saveData();
 			}
 
@@ -159,11 +154,11 @@ public class Shop extends JFrame  {
 		buy4.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				use4.setVisible(true);
-				TestUI.points = Integer.valueOf(points.getText()) - 1000;
+				PlayGame.points = Integer.valueOf(points.getText()) - 1000;
 				points.setText(String.valueOf(Integer.valueOf(points.getText()) - 1000));
 				buy4.setEnabled(false);
 				checkPrice(points);
-				b4IsBought="true";
+				b4IsBought = "true";
 				MainMenu.saveData();
 			}
 		});
@@ -174,7 +169,7 @@ public class Shop extends JFrame  {
 		btnBack.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				setVisible(false);
-				Test.main(null);
+				Main.main(null);
 			}
 		});
 		btnBack.setBounds(231, 585, 142, 42);
@@ -189,7 +184,7 @@ public class Shop extends JFrame  {
 				use3.setEnabled(true);
 				use2.setEnabled(true);
 				use1.setEnabled(true);
-				flag1=4;
+				flag1 = 4;
 			}
 		});
 		use4.setBounds(532, 487, 57, 36);
@@ -204,7 +199,7 @@ public class Shop extends JFrame  {
 				use4.setEnabled(true);
 				use2.setEnabled(true);
 				use1.setEnabled(true);
-				flag1=3;
+				flag1 = 3;
 			}
 		});
 		use3.setBounds(532, 356, 57, 36);
@@ -219,7 +214,7 @@ public class Shop extends JFrame  {
 				use3.setEnabled(true);
 				use4.setEnabled(true);
 				use1.setEnabled(true);
-				flag1=2;
+				flag1 = 2;
 			}
 		});
 		use2.setBounds(532, 224, 57, 36);
@@ -234,7 +229,7 @@ public class Shop extends JFrame  {
 				use3.setEnabled(true);
 				use2.setEnabled(true);
 				use4.setEnabled(true);
-				flag1=1;
+				flag1 = 1;
 			}
 		});
 		use1.setBounds(532, 92, 57, 36);
@@ -242,54 +237,49 @@ public class Shop extends JFrame  {
 
 		this.setVisible(false);
 
-		if(Shop.b1IsBought.equals("true")){
+		if (Shop.b1IsBought.equals("true")) {
 			use1.setVisible(true);
 			buy1.setEnabled(false);
 			checkPrice(points);
-			b1IsBought="true";
-			
+			b1IsBought = "true";
+
 		}
-		if(Shop.b2IsBought.equals("true")){
+		if (Shop.b2IsBought.equals("true")) {
 			use2.setVisible(true);
 			buy2.setEnabled(false);
 			checkPrice(points);
-			b2IsBought="true";
-			
+			b2IsBought = "true";
+
 		}
-		if(Shop.b3IsBought.equals("true")){
+		if (Shop.b3IsBought.equals("true")) {
 			use3.setVisible(true);
 			buy3.setEnabled(false);
 			checkPrice(points);
-			b3IsBought="true";
-			
+			b3IsBought = "true";
+
 		}
-		if(Shop.b4IsBought.equals("true")){
+		if (Shop.b4IsBought.equals("true")) {
 			use4.setVisible(true);
 			buy4.setEnabled(false);
 			checkPrice(points);
-			b4IsBought="true";
-			
+			b4IsBought = "true";
+
 		}
-		
+
 		JLabel lblCoins = new JLabel("New label");
 		lblCoins.setBounds(31, 576, 46, 39);
 		getContentPane().add(lblCoins);
-		
+
 		JLabel lblbackground = new JLabel("New label");
 		lblbackground.setBounds(0, 0, 884, 661);
 		getContentPane().add(lblbackground);
-		
-		
+
 		car1.setIcon(new ImageIcon("pictures\\lambo.png"));
 		car2.setIcon(new ImageIcon("pictures\\lambo2.png"));
 		car3.setIcon(new ImageIcon("pictures\\merc.png"));
 		car4.setIcon(new ImageIcon("pictures\\viper.png"));
 		lblCoins.setIcon(new ImageIcon("pictures\\coins2.png"));
 		lblbackground.setIcon(new ImageIcon("pictures\\garage.jpg"));
-		
-		
-		
 	}
-	
 
 }
